fix: resolve pyright errors, add Bayesian Dirichlet updating, fragmentation module
- pyright 0 errors (19 -> 0): extent tuple cast, FuncFormatter import,
  float() cast, assert narrowing, cf None guard, type: ignore[arg-type]
- bayesian/network.py: _factor_product NumPy broadcasting, Dirichlet CPT methods
- bayesian/water_demand_net.py: calibrate_usage_tier_cpt(), weekday CPT update
- huh_twin/skill_classification.py: Python 3.10 StrEnum backport
- schemas.py: experiment_id pattern ^exp-\d+$, IdeaInput.goal_type default
- entropy/monitor.py: sigma/mu aliases in compute_kl
- gnss/math_utils.py, resilience_twin.py: spoof bias tuning
- models/automation.py, forge.py, report.py, twin/calibrator.py: datetime.UTC fix
- yield_twin/twin.py, gp_surrogate.py: n_candidates threshold, cf None guard
- api/routers/yield_twin.py, model.py: openapi_examples type ignore
- src/fragmentation/: growth fragmentation equation module (eigenanalysis, kernels, ML)
- configs/model_registry/growth_fragmentation.yaml: model registry entry
- tests/test_fragmentation_*.py: eigenanalysis/kernels/ML test suites

+++ src/twin/core.py
@@ -494,7 +494,7 @@ def _mean_clustering(W: np.ndarray, thresh: float = 0.30) -> float:
         e = sum(A[a, b] for a in nbrs for b in nbrs if a < b)
         cc_sum += e / (d_i * (d_i - 1) / 2)
         n_counted += 1
-    return cc_sum / max(n_counted, 1)
+    return float(cc_sum / max(n_counted, 1))
 
 
 class StructDepMonitor:
